+++ packages/nimble-blazor/NimbleBlazor/EventHandlers.cs
@@ -1,4 +1,5 @@
 ﻿using System.Globalization;
+using System.Text.Json.Serialization;
 using Microsoft.AspNetCore.Components;
 
 namespace NimbleBlazor;
@@ -38,6 +39,29 @@ public class TableRowSelectionEventArgs : EventArgs
     public IEnumerable<string>? SelectedRecordIds { get; set; }
 }
 
+public class TableColumnConfigurationEventArgs : EventArgs
+{
+    public IEnumerable<TableColumnConfiguration>? Columns { get; set; }
+}
+
+public class TableColumnConfiguration
+{
+    public string? ColumnId { get; set; }
+
+    public int? SortIndex { get; set; }
+
+    [JsonConverter(typeof(TableColumnSortDirectionConverter))]
+    public TableColumnSortDirection SortDirection { get; set; }
+
+    public int? GroupIndex { get; set; }
+
+    public bool Hidden { get; set; }
+
+    public double FractionalWidth { get; set; }
+
+    public double? PixelWidth { get; set; }
+}
+
 [EventHandler("onnimbletabsactiveidchange", typeof(TabsChangeEventArgs), enableStopPropagation: true, enablePreventDefault: true)]
 [EventHandler("onnimblecheckedchange", typeof(CheckboxChangeEventArgs), enableStopPropagation: true, enablePreventDefault: true)]
 [EventHandler("onnimblemenubuttontoggle", typeof(MenuButtonToggleEventArgs), enableStopPropagation: true, enablePreventDefault: false)]
@@ -46,6 +70,7 @@ public class TableRowSelectionEventArgs : EventArgs
 [EventHandler("onnimbleactionmenutoggle", typeof(TableActionMenuToggleEventArgs), enableStopPropagation: true, enablePreventDefault: false)]
 [EventHandler("onnimbleactionmenubeforetoggle", typeof(TableActionMenuToggleEventArgs), enableStopPropagation: true, enablePreventDefault: false)]
 [EventHandler("onnimbletablerowselectionchange", typeof(TableRowSelectionEventArgs), enableStopPropagation: true, enablePreventDefault: false)]
+[EventHandler("onnimbletablecolumnconfigurationchange", typeof(TableColumnConfigurationEventArgs), enableStopPropagation: true, enablePreventDefault: false)]
 public static class EventHandlers
 {
 }